+++ asapdiscovery-data/asapdiscovery/data/selectors/mcs_selector.py
@@ -7,7 +7,7 @@ from asapdiscovery.data.schema_v2.complex import Complex, ComplexBase, PreppedCo
 from asapdiscovery.data.schema_v2.ligand import Ligand
 from asapdiscovery.data.schema_v2.pairs import CompoundStructurePair
 from asapdiscovery.data.selectors.selector import SelectorBase
-from asapdiscovery.docking.docking_v2 import DockingInputPair
+from asapdiscovery.docking.docking import DockingInputPair
 from pydantic import Field
 
 logger = logging.getLogger(__name__)
