+++ asapdiscovery-docking/asapdiscovery/docking/tests/test_scorers.py
@@ -1,5 +1,5 @@
 import pytest
-from asapdiscovery.docking.scorer_v2 import (
+from asapdiscovery.docking.scorer import (
     ChemGauss4Scorer,
     GATScorer,
     MetaScorer,
