+++ asapdiscovery-simulation/asapdiscovery/simulation/simulate.py
@@ -11,7 +11,7 @@ import pandas as pd
 from asapdiscovery.data.dask_utils import actualise_dask_delayed_iterable
 from asapdiscovery.data.enum import StringEnum
 from asapdiscovery.data.openeye import save_openeye_pdb
-from asapdiscovery.docking.docking_v2 import DockingResult
+from asapdiscovery.docking.docking import DockingResult
 from mdtraj.core.residue_names import _SOLVENT_TYPES
 from mdtraj.reporters import XTCReporter
 from openff.toolkit.topology import Molecule
